Modification: Execute project migration tasks before login
Modification: Catch exception in executeProjectTasks

Modification: RemoveRoutineFolderTask in localRepository is executed only if project is local

git-svn-id: http://talendforge.org/svn/tos/trunk@1903 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.repository.localprovider/src/main/java/org/talend/repository/localprovider/model/migration/RemoveRoutineFolderTask.java
@@ -45,9 +45,14 @@ public class RemoveRoutineFolderTask extends AbstractMigrationTask implements IP
      * @see org.talend.core.model.migration.IMigrationTask#execute()
      */
     public boolean execute(Project project) {
+        if (!project.isLocal()) {
+            return true;
+        }
+
         try {
             IProject iproject = ResourceModelUtils.getProject(project);
-            String oldRoutinesPath = "routines"; // Routines path as it was in talend v1.0.n and until 1.1.m2 //$NON-NLS-1$
+            String oldRoutinesPath = "routines"; // Routines path as it was in talend v1.0.n and until 1.1.m2
+            // //$NON-NLS-1$
             IFolder f2 = ResourceUtils.getFolder(iproject, oldRoutinesPath, false);
             ResourceUtils.deleteFolder(f2);
             return true;
